+++ source/elm.ts
@@ -1,0 +1,7 @@
+import * as type from "./type";
+
+export const generateCode = (
+  customTypeDictionary: ReadonlyMap<string, type.CustomType>
+): string => {
+  return "";
+};
