+++ test/test.ts
@@ -2,27 +2,27 @@ import * as t from "../source/index";
 import * as generator from "js-ts-code-generator";
 
 describe("test", () => {
-  const data: ReadonlyMap<string, t.CustomType> = new Map([
+  const data: ReadonlyMap<string, t.type.CustomType> = new Map([
     [
       "",
       {
         description: "サンプル",
-        body: t.customTypeBodyProduct([
+        body: t.type.customTypeBodyProduct([
           {
             name: "UInt32",
             description: "0～4294967295 32bit符号なし整数",
-            parameter: t.tagParameterNothing
+            parameter: t.type.tagParameterNothing
           },
           {
             name: "String",
             description: "文字列",
-            parameter: t.tagParameterNothing
+            parameter: t.type.tagParameterNothing
           }
         ])
       }
     ]
   ]);
-  const code: generator.Code = t.generateTypeScriptCode(data);
+  const code: generator.Code = t.typeScript.generateCode(data);
   const codeAsString = generator.toNodeJsOrBrowserCodeAsTypeScript(code);
   it("return string", () => {
     console.log(codeAsString);
